Let the python-twitter library handle message splitting
Twitter auto-converts urls to twitter shortened URLs. This can affect
the math for message splitting. The python-twitter library has this
implemented already, including all of the logic necessary ... so just
use it.

The character '\u2026' is the unicode horizontal ellipsis and is used
to indicate that the message is continued.

Change-Id: I5492a2792f5d11dccbe4a3454396262341449eab